src/kafka/requirements.txt
avro==1.11.1
certifi==2022.12.7
confluent-kafka==2.0.2
elastic-apm==6.15.1
elastic-transport==8.4.0
elasticsearch==8.6.2
python-dotenv==1.0.0
urllib3==1.26.14
wrapt==1.15.0
